Revert "Some optimization on Alpha-beta"
This reverts commit bc6c11fe3832ea7fcddeb3f5d449c8cc6df7d0bb.

# Conflicts:
#	AI_Min_Max.cpp
#	AI_Min_Max.h
#	GameScene.cpp

+++ Player.h
@@ -21,7 +21,7 @@ class Player{
 public:
     string name;
     virtual vector<Move> genAllMoves( Board &board);
-    virtual int eval( Board &board);
+    virtual float eval( Board &board);
     vector<Move> genAMove( Board &board, Pieces * fromPiece);
     
 };
